Moves entities into their own file for organization's sake.
